Fixed issue where used Float instead of Double

+++ GaslightVillage/src/IOwner.java
@@ -1,5 +1,5 @@
 public interface IOwner {
   public void addAssessment(IProject project);
-  public Float calculateNewFee();
+  public Double calculateNewFee();
 
 }
